+++ src/core/ioremap.cc
@@ -40,7 +40,7 @@ bitmap_t create_bitmap(uint64_t start, int ele_size, int bit_nr) {
         .data = nullptr,
         .start = start,
         .ele_size = ele_size,
-        .bit_nr = bit_nr,
+        .bit_nr = (uint64_t)bit_nr,
     };
 
     b.data = (uint64_t *)kmalloc(bit_nr / BITS_PER_BYTE);
